LoadEntities delegates batches to async job
Instead of chewing through every batch in the course of running
LoadEntities, put the ids for each batch into an async job and let that
job put together a fat payload of all the import events for all those
ids - less network, and atomic retries (because it's all or nothing).

Remove 'sleep' option because it's unnecessary - risk of overwhelming
the queue is much smaller.