Add .tool-versions
Because the async gem, which is a (distant) dependency of
github_changelog_generator, introduces a Ruby version constraint of
>=3.1.1 for the 2.x series, bundling on a machine with Ruby 3.1.1
installed will result in a different Gemfile.lock than bundling on a
machine with an earlier Ruby version. Note than this is a development
dependency only so the official supported ruby version is still 2.7.

In fact the behaviour of bundler here is quite confusing because it says
that async 2.x.x has been *withdrawn* by the author, which isn't
actually true — it's just that the version in the Gemfile.lock can't be
reconciled with the Ruby version you're running. They've already fixed
it (but haven't released it yet afaics).

https://github.com/rubygems/rubygems/pull/5525

We use asdf everywhere else - use it here too