Remove invalid UTF-8 from request headers
Users are able to craft request content which goes to BigQuery by
setting request headers. This can cause issues at the point of job
serialization (Sidekiq can't serialize the args) or at the point of
sending to BQ (the Google Cloud library can't build the JSON), when
there's a string present that's not valid UTF-8.

We could preemptively rountrip events to JSON and back again before
putting them on the queue, but the Google library expects hashes, so we
can't reuse that JSON. Instead, focus on the things we know can be
malformed and fix those: ua, referer and path. We don't need to worry
about query params as Rails will reject these out of hand - see
ActionDispatch::Request::Utils.check_param_encoding).

We decided to still send these requests, because we send everything,
but with � in place of unparseable bytes (which is how String#scrub
replaces them).