handle data that have JSON-friendly representation
When converting event data into the JSON payload to send to BQ, have
special handling for Hash objects. However, some data will come in as on
object that isn't immediately convertable to something JSON-friendly,
such as models (e.g. StoreModel or ActiveModel objects).

This change converts everything to a more standard JSON-friendly
representation before processing for sending to BQ, meaning that models
are converted to a Hash object which we can then to_json as needed.