detect private current_* methods
Some services may declare current_user (or current_* really) as private
methods if there is no need to expose that method in the controller.
Previously we would miss out on sending the user_id with the event if
this was the case, this change allows us to detect these as private
methods.

For testing purposes, we've made the current_* methods private since
that ensures we work with both private and public versions of these
methods.