Handle models which share a table
When apps use single-table inheritance, multiple Rails models
share the same database table - i.e. they are considered a single
'entity' by DfE Analytics.

This isn't good, because it means that at most one model can be
associated with a single table. In turn this means that we will not know
to attach callbacks to the other models in the STI heirarchy. In
practice, due to the order in which ActiveRecord::Base.descendants
returns the models, this seems to be the alphabetically first child, not
the parent, and not any of the siblings.

Treating the right hand side of the mapping as a list of models, not a
single model, allows us to handle this. We have to update the code in
a couple places to expect arrays instead of strings.