Do not attempt initialization when the user is changing the database
db:* rake commands are often useful to bring the database into line with
what DfE::Analytics expects, so it can be painful not to be able to run
them when the config is inconsistent. Allow these commands at all times -
full checks will run on app or test boot anyway.