Prevent duplicate fields in blocklist when using STI
When the application has models that are configured with single table
inheritance the generated blocklist will contain duplicate field names for
the table (once for each model).

This is due to the behaviour of `attribute_names` in Rails; it retrieves
the names directly from the table columns and doesn't filter based on the
context of the subclass.

To avoid this we need to `uniq` the attribute names generated for the
blocklist.

Example blocklist prior to change (assuming a parent/child using STI):

```
{:with_model_candidates_43990_9360=>["id", "first_name", "id", "first_name"]}
```

And after change:

```
{:with_model_candidates_43990_9360=>["id", "first_name"]}
```