Add publishing workflow
This flow allows tags to be built into gems and those gems published to
GitHub's container registry, simplifying the Gemfile entries for
projects that consume it.

The Ruby and Rails versions align with the lowest versions in the test
matrix.

I didn't add Rubygems publishing because I'm not sure who's account it
should be added to, perhaps that could follow.