Created a failing spec
While investigating an issue of missing data we started investigating the implementation of sending events while inside Rails transactions.

This commit reproduces a bug that may be contributing to some of our data issues.
The existing spec assets that no events are send if the transaction is rolled back, however the spec was passing due to `allowlist_fields` not being set.
After setting the `allowlist_fields` we see a test failure as the `SendEvents` job is called.